+++ prog/weave-client/main.go
@@ -1,27 +1,60 @@
 package main
 
 import (
-	"fmt"
-	"os"
-	"os/exec"
+	"crypto/tls"
 	"encoding/json"
 	"io"
 	"net/http"
-	"crypto/tls"
+	"os"
+	"os/exec"
 	"strings"
 
-	logrus "github.com/sirupsen/logrus"
 	util "github.com/AVENTER-UG/util/util"
 	cfg "github.com/AVENTER-UG/weave/prog/weave-client/types"
+	logrus "github.com/sirupsen/logrus"
 )
 
 var config cfg.Config
 
 func init() {
-	config.APIURL = util.Getenv("WEAVE_BACKEND_API", "root")
+	config.APIHost = util.Getenv("WEAVE_BACKEND_API_HOST", "root")
 	config.APIToken = util.Getenv("WEAVE_BACKEND_API_TOKEN", "")
 	config.SSL = stringToBool(util.Getenv("WEAVE_BACKEND_SSL", "false"))
 
+	setWeaveConfiguration()
+}
+
+// setWeaveConfiguration get the configuration from backen and set
+// the config for weaver
+func setWeaveConfiguration() {
+	body, err := httpRequest("GET", "/api/weave/v0/network")
+
+	if err != nil {
+		logrus.WithField("func", "weave-client.getWeaveConfiguration").Error(err.Error())
+		return
+	}
+
+	var net cfg.WeaveNetwork
+	if err := json.Unmarshal(body, &net); err != nil {
+		panic(err)
+	}
+
+	// read all peers
+	var peers strings.Builder
+	for i, p := range net.Peers {
+		if i > 0 {
+			peers.WriteByte(' ')
+		}
+		peers.WriteString(p.Hostname)
+	}
+
+	os.Setenv("WEAVE_PASSWORD", net.EncryptionPassword)
+	os.Setenv("IPALLOC_RANGE", net.OverlayCIDR)
+	os.Setenv("WEAVE_PEERS=", peers.String())
+}
+
+// httpRequest is the global function to connect with the backend
+func httpRequest(method, url string) ([]byte, error) {
 	protocol := "https"
 	if !config.SSL {
 		protocol = "http"
@@ -33,7 +66,7 @@ func init() {
 		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
 	}
 
-	req, err := http.NewRequest("GET", protocol+"://"+config.APIURL, nil)
+	req, err := http.NewRequest(method, protocol+"://"+config.APIHost+url, nil)
 	if len(config.APIToken) > 0 {
 		req.Header.Set("Authorization", config.APIToken)
 	}
@@ -42,25 +75,17 @@ func init() {
 
 	if err != nil {
 		logrus.WithField("func", "init").Error("Could not connect to backend: ", err.Error())
-		panic(err)
+		return nil, err
 	}
 
 	if res.StatusCode == http.StatusOK {
 		defer res.Body.Close()
 	}
 
-	body, _ := io.ReadAll(res.Body)
-
-	// TODO
-	var net cfg.WeaveNetwork
-	if err := json.Unmarshal(body, &net); err != nil {
-		panic(err)
-	}
-
-	fmt.Printf("Weave Network: %+v\n", net)
+	return io.ReadAll(res.Body)
 }
 
-
+// stringToBool convert true/false string to boolean
 func stringToBool(par string) bool {
 	return strings.Compare(par, "true") == 0
 }